rename keys to crypto, comment funcs
