+++ crypto/ed25519_test.go
@@ -1,4 +1,4 @@
-package keys
+package crypto
 
 import (
 	"reflect"
